Created interface for PCES writer.
